SSAFY_PEDIA_Django_Back/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
